fix bug: test not run due to missing billed_items directory
